The CPU only supports 32-bit physical addresses. No point in keeping the result of translated addresses 64-bit.

+++ DotN64/CPU/VR4300/CP0/VR4300.SystemControlUnit.cs
@@ -27,14 +27,14 @@ namespace DotN64.CPU
             /// Translates a virtual address into a physical address.
             /// See: datasheet#5.2.4 Table 5-3.
             /// </summary>
-            public ulong Translate(ulong address)
+            public uint Translate(ulong address)
             {
                 switch (address >> 29 & 0b111)
                 {
                     case 0b100: // kseg0.
-                        return address - 0xFFFFFFFF80000000;
+                        return (uint)(address - 0xFFFFFFFF80000000);
                     case 0b101: // kseg1.
-                        return address - 0xFFFFFFFFA0000000;
+                        return (uint)(address - 0xFFFFFFFFA0000000);
                     default:
                         throw new Exception($"Unknown memory map segment for location 0x{address:X16}.");
                 }
